removed version and encode_kwargs from task_result
